Move file
active_record_ext/connection_adapters/schema_statements.rb -> active_record_ext.rb
