+++ lib/activerecord/simple_index_name.rb
@@ -9,5 +9,5 @@ end
 if defined?(Rails)
   require "activerecord/simple_index_name/railtie"
 else
-  require "activerecord/simple_index_name/active_record_ext/connection_adapters/schema_statements"
+  require "activerecord/simple_index_name/active_record_ext"
 end
