+++ lib/activerecord/simple_index_name/active_record_ext.rb
