+++ lib/activerecord/simple_index_name/railtie.rb
@@ -2,7 +2,7 @@ module Activerecord
   module SimpleIndexName
     class Railtie < ::Rails::Railtie
       on_load :active_record do
-        load "#{__dir__}/../simple_index_name/active_record_ext/connection_adapters/schema_statements.rb"
+        load "#{__dir__}/active_record_ext.rb"
       end
     end
   end
